Allow a default menu-choice to be read
When the system boots control eventually reaches the monitor, which
presents a menu, and reads input to decide what to do.

This pull-request configures things such that the file `/CPM/START.CFG`
will be read, if present, and the *first character* from that file
will be saved into a one-byte input-buffer.

If the input-buffer is non-zero it will be used in preference for user
input:

* The net result is you should be able to boot to CPM via
  * `echo c > /path/to/mount/CPM/START.cfg`

The input-buffer is only a single byte, but can be used to select
(c)pm, (t)iny basic, or any of the other options.